% Generated by build script
\name{boe_millennium}
\alias{boe_millennium}
\title{Download BoE Millennium macroeconomic dataset}
\usage{
boe_millennium()
}
\arguments{
\item{sheet}{Sheet name or number to read. NULL lists all sheets.}
}
\value{tibble from the specified sheet, or tibble of sheet names if NULL}
\description{Download BoE Millennium macroeconomic dataset}
