% Generated by build script
\name{boe_yield_curve}
\alias{boe_yield_curve}
\title{Latest UK government yield curve data (daily)}
\usage{
boe_yield_curve()
}
\arguments{
\item{curve}{Which curve: "nominal" (default), "real", "inflation", "ois"}
}
\value{tibble: date + maturity columns with yield values}
\description{Latest UK government yield curve data (daily)}
